+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
 
 - new abstact class AbstractMutableComboBoxModel for a generic combobox model that can add and remove elements from the combobox dynamically
 - new decorator class for JCheckBox that encapsulate a model and can be changed the checked value over the model
+- new class StringMutableComboBoxModel created that extends the AbstractMutableComboBoxModel
 
 Version 5.8
 -------------